requirements.txt
numpy==1.11.3
scipy==1.0.0
pandas==0.22.0
matplotlib==2.0.0
six==1.11.0
python-telegram-bot